Add files via upload
Series Data Structure in Pandas
Pandas Series is a one-dimensional labeled data structure that can hold data of any type (integer, float, string, etc.).
It can be created using a Python list, tuple, dictionary, NumPy array, or a scalar value. However, it cannot be created using a Python set, as sets are unordered, and Pandas raises an error when you try to create a Series from one.
In the above file code for same is given you can check out.