Add files via upload
DataFrame Data Structure
Pandas DataFrame is a two-dimensional labeled data structure with rows and columns, similar to a table. It can be created using the built-in DataFrame() function from various data types such as a list, tuple, dictionary, nested list of sets, or even from a Series using the to_frame() method.the code for same is provided in the below file please check out.